[MODFISTO-273] - Refactoring
